% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_simulate.R
\name{sim_ar}
\alias{sim_ar}
\title{Simulating from AR model}
\usage{
sim_ar(
  ar_order,
  length,
  params = NULL,
  sample_series = NULL,
  no_series = 1,
  max_root = 5,
  burn_in = 100
)
}
\arguments{
\item{ar_order}{The order of the AR model to be used for simulation}

\item{length}{The length of the simulated series}
}
\value{
A list with named arguments `parameters` and the `series` generated
}
\description{
This function simulates a defined number of series from a defined AR model.
}
