% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssimulate-package.R
\docType{package}
\name{tssimulate-package}
\alias{tssimulate}
\alias{tssimulate-package}
\title{tssimulate: Simulating Time Series from Data Generating Processes}
\description{
A package for simulating time series from a number of data generating processes including AR, SAR, Chaotic Logistic Map, Mackey-Glass Equations and Fourier Terms with given options from the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HansikaPH/tssimulate}
  \item Report bugs at \url{https://github.com/HansikaPH/tssimulate/issues}
}

}
\author{
\strong{Maintainer}: Hansika Hewamalage \email{hansika.hewamalage@monash.edu}

Authors:
\itemize{
  \item Christoph Bergmeir \email{christoph.bergmeir@monash.edu}
  \item Kasun Bandara \email{kasun.bandara@unimelb.edu.au}
}

}
\keyword{internal}
